+++ src/compiler/compiler.ts
@@ -32,6 +32,14 @@ export class Compiler implements d.Compiler {
       if (config.sys.semver.prerelease(config.sys.compiler.version)) {
         config.logger.warn(config.sys.color.yellow(`This is a prerelease build, undocumented changes might happen at any time. Technical support is not available for prereleases, but any assistance testing is appreciated.`));
       }
+      if (config.devMode && config.buildEs5) {
+        config.logger.warn(`Generating ES5 during development is a very task expensive, initial and incremental builds will be much slower. Drop the '--es5' flag and use a modern browser for development.
+        If you need ESM output, use the '--esm' flag instead.`);
+      }
+      if (config.devMode && !config.enableCache) {
+        config.logger.warn(`Disabling cache during development will slow down incremental builds.`);
+
+      }
       config.logger.debug(`${details.platform}, ${details.cpuModel}, cpus: ${details.cpus}`);
       config.logger.debug(`${details.runtime} ${details.runtimeVersion}`);